clean: node-xapian.h -  created a single header file
